requirements.txt
streamlit>=1.28.0
requests>=2.28.0
PyMuPDF>=1.23.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
openai>=1.0.0
huggingface_hub>=0.16.0
langchain>=0.0.300
numpy>=1.24.0
pandas>=1.5.0
python-dotenv>=1.0.0

# LangChain Extended Components
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-huggingface>=0.0.1

# Text Processing & NLP
nltk>=3.8.1
tiktoken>=0.5.0
transformers>=4.30.0

# ML/AI Core
torch>=2.0.0
scikit-learn>=1.3.0

# Utility Libraries
pydantic>=2.0.0
typing-extensions>=4.7.0
tqdm>=4.65.0

# HTTP Client (better than requests for async)
httpx>=0.24.0